Fixed overflow issue in "simple" example

+++ examples/simple/simple.ino
@@ -15,7 +15,7 @@ ASB asb0(0x123);
 //Start new CAN-Bus with 125KBps and 8MHz crystal using CS pin 10 and Interrupt pin 2
 ASB_CAN asbCan0(10, CAN_125KBPS, MCP_8MHz, 2);
 
-unsigned int lastSend;
+unsigned long lastSend;
 
 void setup() {
   //Initialize Serial port
